Extract registry loading and validation
