+++ src/binary/constants.ts
