+++ src/binary/ibc-setup/commands/init.ts
@@ -1,16 +1,14 @@
 import fs from 'fs';
-import os from 'os';
 import path from 'path';
 
-import Ajv, { JSONSchemaType } from 'ajv';
 import axios from 'axios';
 import yaml from 'js-yaml';
 
-import { appFile, registryFile } from '../../../constants';
-import { Registry } from '../types';
+import { appFile, registryFile } from '../../constants';
 import { deriveAddress } from '../utils/derive-address';
 import { generateMnemonic } from '../utils/generate-mnemonic';
 import { getDefaultHomePath } from '../utils/get-default-home-path';
+import { loadAndValidateRegistry } from '../utils/load-and-validate-registry';
 import { resolveRequiredOption } from '../utils/resolve-required-option';
 
 export type Options = {
@@ -61,48 +59,7 @@ export async function run(options: Options) {
     throw new Error(`${registryFilePath} must be a file.`);
   }
 
-  const registry = yaml.load(fs.readFileSync(registryFilePath, 'utf-8'));
-
-  const ajv = new Ajv({ allErrors: true });
-  const schema: JSONSchemaType<Registry> = {
-    type: 'object',
-    required: ['chains', 'version'],
-    additionalProperties: false,
-    properties: {
-      version: {
-        type: 'number',
-      },
-      chains: {
-        type: 'object',
-        minProperties: 2,
-        required: [],
-        additionalProperties: false,
-        patternProperties: {
-          '^(.*)$': {
-            type: 'object',
-            required: ['chain_id', 'gas_price', 'hd_path', 'prefix', 'rpc'],
-            additionalProperties: false,
-            properties: {
-              chain_id: { type: 'string' },
-              prefix: { type: 'string' },
-              gas_price: { type: 'string' },
-              hd_path: { type: 'string' },
-              rpc: { type: 'array', items: { type: 'string' } },
-            },
-          },
-        },
-      },
-    },
-  };
-  const validate = ajv.compile(schema);
-  if (!validate(registry)) {
-    const errors = (validate.errors ?? []).map(
-      ({ dataPath, message }) => `"${dataPath}" ${message}`
-    );
-    throw new Error(
-      [`${registryFile} validation failed.`, ...errors].join(os.EOL)
-    );
-  }
+  const registry = loadAndValidateRegistry(registryFilePath);
 
   const [chainSrc, chainDest] = [options.src, options.dest].map((chain) => {
     const chainData = registry.chains[chain];
